octofit-tracker/backend/requirements.txt
Django==4.1
djangorestframework==3.14.0
pymongo==4.3.3
django-cors-headers==3.14.0
python-dotenv==1.0.0